chore(Javascript): add DFS & BFS algorithm
